requirements.txt
discord=2.3.2
toml=0.10.2
python-dateutil=2.8.2